requirements.txt
flask~=2.2.2
pillow~=9.4.0
torch~=1.13.1
torchvision~=0.14.1
torchaudio
numpy~=1.24.1
requests~=2.28.2
opencv-python~=4.7.0.68